Added testing of the fvs[Get|Set]SpeciesAttrs code.
Fixed a bug in how stoppoint 7 works. Added more code to the test of the
APIviaR that would have shown the error had the test been added prior to
now.  Updated the 'save' files to account for these changes.

Added test for using fvsCutNow() option in rFVS and added code and
tests that trap using the thinprsc keyword in a way that causes a conflict
with the use of the rFVS fvsCutNow function.